Build innie for Scala 2.13
This required me to fix some code to remove warnings about the use of
Either.right (Either is right-biased in Scala 2.13), as well as
adjusting some messages from the Scala parser combinator library (in
Scala 2.13, token names are in the format 'NAME' in error messages,
whereas they used to be in the format `NAME').

I also had to remove support for Scala 2.10 and 2.11; Either.getOrElse
does not exist in those versions.